requirements.txt
peewee
nicegui
pyyaml